add azure pipelines for testing (#11)
This adds two azure pipelines. One runs the service tests and integration tests
using the checked-in, pre-built enclave library. The other rebuilds the enclave
before running the tests.

In order to support both possible enclave library locations, we had to give the integration tests some help figuring out which to use. Currently, the default checked in enclave is hardcoded in the integration tests but looks for a `current.mrenclave` resource to override the desired enclave filename. (Maybe instead we should check `current.mrenclave` file? open to feed back)

We cache the maven resources in both pipelines and the llvm-bolt binary in the
enclave rebuilding one.

You can see the pipelines in action at
https://dev.azure.com/signal-testing/directory-testing/_build