turn JNI code into rust (#24)
turn JNI code into rust

This is a backwards-compatible rewrite of the JNI bindings into Rust.

This patch is a combination of Jeffrey and I's work. Jeffrey did the
good stuff; the bad stuff is all mine.

The majority of the core rewrite work is in `jniimpl.rs`. It implements
the JNI C API found by generating the Java JNI header for SgxEnclave
(see the `-h` in the pom.xml) and running `jbindgen` on that header.

A new `enclave-ffi-sdk` project was added. It has enough differences
with `sgx_common` that we can resolve them later -- mostly by improving
`sgx_common`'s SgxResult and similar (e.g. SgxStatus is not great as an
error type since it includes a success state). Since that implied KBS
work and a separate patch for sgx_common, I've skipped it.

The `cargo` layout here is not ideal. Preferably, we'd move the Cargo
workspace up to the top-level instead of having one deep in `service`.
I'm very open to moving it up. I just didn't want to take the extra
cargo churn. (Perhaps it wouldn't be too bad.)

Fixes CDS-44