frontend: correct phone and uuid buffers race (#50)
frontend: correct phone and uuid buffers race

This patch is fixing a race between modifications to the phone and
uuid buffers and the reading of those buffers in the Rust JNI and
enclave code.

We do so by implementing double-buffering and explicitly "borrowing"
the read-only buffers under a read-lock. A double-buffered map means
users will not see new contacts in the lookup functions until the map's
`update` function swaps the read and write buffers. We schedule an
`update` for every 30 seconds and the double-buffered map will not swap
if there have been no writes since the last `update`.

We double our memory allocations, and writes will contend with buffer
swaps as the map currently copies its entire freshly swapped read buffer
back to the write buffer.

DirectoryHashSet deserves a rename, and DirectoryManager could use some
of its responsibilities split up, but this is already touching a core
part of the system, I didn't want more work to go in.

Fixes CDS-42