enclave: move benchmarks to separate crate
When benchmarking the enclave, we want to benchmark the "bolt
mitigated" code.  Unfortunately, the enclave shared library is not
directly usable by the benchmarking code because most global symbols
have been removed by the linking process.

This patch reworks the Makefile benchmark build to create a new shared
library and crate, which is separate from the enclave shared library
and crate.  Both crates share a common static library.

- the cds_enclave crate remains a no-std staticlib crate, generating
  libcds_enclave.a

- the production enclave (the shared library libcds_enclave.so) is
  built from libcds_enclave.a.  Bolt operates on the libcds_enclave.so.

- the Makefile creates a new shared library, libcds_benchmark.so,
  which is built from the same static lib as the production enclave.
  The only difference is this shared library exports symbols that the
  benchmarking code needs.  This shared library is also processed by
  bolt.