add hand-hacked assembly for the LVI mitigations (#83)
This is the outcome of a long, whole-team effort to get the LVI
mitigations to be speedy enough for us to use in CDS. CDS was only able
to do single digit number of requests before this.

We expose a new hot-loop function, tell BOLT to skip mitigating it, and
instead hack the assembly directly to get the lfences into the right
place. The lvi_checker tool also had to learn about skipping functions.

Along the way, we have to now be very careful about updating
cds-enclave-hash.rs and its accompanying assembly file. The assembly is
no longer rebuilt on each build, because of the by-hand additions.