frontend: add failure callbacks inside JNI (#88)
Using the magic of CompletableFuture, we add support for routing SGX
errors back up from the nativeServerStop (which is the JNI call that
actually processes the data in a request).

Fortunately, we had one nearby that just needed to be slid into place.
That lets us remove the now unused callback class and is pretty.

There are still some eprintln! in the JNI code where the JNI
interactions themselves fail inside the enclave's thread. There, of
course, isn't any other mechanism to contact the Java code that's
waiting and trying to tap into, say, the Java metrics or what have you,
will likely not work for the same reason the JNI code had already
failed.