service: generate unique hostId and check in request envelopeMap
When the server starts generate a unique hostId.

In the attestation response, this hostId is sent back to the client.
The subsequent directory request from the client must include this
hostId.

This is to help verify that the clients are hitting the same server
for both attestation and the discovery request.