service: fix flaky testReconcileRange() test
The test sometimes dies expecting to see at least one jedis.publish()
call.

Previously, the logic of the test never leads to a jedis.publish().

However, *most times* a jedis.publish() would occur because the
KeepAliveSender class creates a jedis.publish() on a background
thread.

This fix here adds a when() clause for when the expected
directoryCache.userAdd(<UUID, phoneNumber>) occurs, leading ultimately
to the jedis.publish() call.