Convert all SgxExceptions for a server call (#100)
Server calls can either complete immediately or through a
future. SgxExceptions were only being converted if the exception
occurred immediately. If the exception occurred and was posted to the
future directly from the JNI code, it was not being converted. This
change ensures the CompletionStage translates the exception regardless
of where the exception gets posted to the future from.