Move execution of the borrow on DirectoryMap into the Rust code
Allocating a Vec<u8> and treating it like a Vec<u64> and Vec<u128>
isn't quite correct due to potential alignment issues. Rather than try
to solve for the alignment issues while keeping the type at a byte
level, this pushes the borrow function into Rust so the type
information is not lost with a pass through Java.

This additionally allows for removing the weird cruftiness that was
taking place in the server 'stop' function in sgx_enclave.rs