Reduce memory allocation in directory map native
When passed a very large buffer to read from or write into, the
current implementation in the native directory map would attempt to
allocate a similarly large byte buffer in the JVM. What is actually
needed is just a large enough byte buffer to amortize the cost of the
read call across enough bytes inside a loop. This should substantially
reduce the memory footprint inside the JVM of the read from and write
to calls on this native directory map.